fix:move sql.go file to new position
